ethercatmcAxis.cpp: Don't call initialPoll() in constructor
Fix a race condition when the first poll is done towards an axis:
Before this commit, the initialPoll() was done in the constructor
of an axis.
The constructor was run from the "main" thread, the IOC shell.
The initialPoll() needs to know, which features the axis has,
especially it needs to know if it is a simulator, to handle lines like

simOnly setADRdouble  0x5000 0xD 14

correctly.
However, the features are common for all axes, so that the controller
reads them in another thread, the poll thread.
When the poll thread starts slow, we don't have any features,
and the simulated axis does not get any soft limits (and TC090 fails).

Fis it by removing the initialPoll() from the constructor.